+++ CHANGELOG.md
@@ -3,6 +3,6 @@
 ## LDAPCP v11 enhancements & bug-fixes
 
 * Fixed no result returned under high load, caused by a thread safety issue where the same filter was used in all threads regardless of the actual input
-* Improved validation of changes done to ClaimTypes collection
+* Improved validation of changes made to ClaimTypes collection
 * Added method ClaimTypeConfigCollection.GetByClaimType()
 * Implemented unit tests